Update imports due to directories refactoring
