+++ app/src/main/res/layout/activity_user_dashboard.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".User_Dashboard">
+    tools:context=".UserDashboard">
 
     <Button
         android:id="@+id/buttonViewUnits"
